+++ CHANGELOG.md
@@ -14,6 +14,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * [mse214830: Selecting text in profile activity comments causes unexpected clipping](http://meta.stackexchange.com/q/214830) (CSS)
 * [mse230392: Layout bug while viewing vote count in Meta Stackexchange](http://meta.stackexchange.com/q/230392) (CSS)
 * [mse230370: The entire answer is moving when hovering the bounty button](http://meta.stackexchange.com/q/230370) (CSS, meta.SE & academia.SE only)
+* [english4719: Background of long OP username looks ugly in comments](http://meta.english.stackexchange.com/q/4719) (CSS, english.SE only)
 
 **Other changes:**